worker/requirements-test.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-httpx>=0.30.0
pytest-cov>=4.0.0
python-dotenv>=1.0.0
lxml>=5.0.0
